+++ super-agent/src/lib.rs
@@ -8,6 +8,8 @@ pub mod sub_agent;
 pub mod super_agent;
 pub mod utils;
 
+pub mod values;
+
 #[cfg(feature = "k8s")]
 pub mod k8s;
