+++ super-agent/src/sub_agent.rs
@@ -6,7 +6,6 @@ pub mod event_processor;
 pub mod event_processor_builder;
 pub mod health;
 pub mod persister;
-pub mod values;
 
 #[cfg(feature = "k8s")]
 pub mod k8s;
