+++ super-agent/src/sub_agent/error.rs
@@ -1,11 +1,11 @@
 use super::effective_agents_assembler::EffectiveAgentsAssemblerError;
-use super::values::values_repository::ValuesRepositoryError;
 use crate::agent_type::agent_values::AgentValuesError;
 use crate::event::channel::EventPublisherError;
 use crate::opamp::client_builder::OpAMPClientBuilderError;
 use crate::opamp::hash_repository::repository::HashRepositoryError;
 use crate::opamp::remote_config::RemoteConfigError;
 use crate::super_agent::config::SuperAgentConfigError;
+use crate::values::values_repository::ValuesRepositoryError;
 use opamp_client::StartedClientError;
 use opamp_client::{ClientError, NotStartedClientError};
 use std::time::SystemTimeError;
