+++ super-agent/src/sub_agent/event_processor.rs
@@ -5,9 +5,9 @@ use crate::opamp::hash_repository::HashRepository;
 use crate::opamp::operations::stop_opamp_client;
 use crate::sub_agent::error::SubAgentError;
 use crate::sub_agent::health::with_start_time::HealthWithStartTime;
-use crate::sub_agent::values::values_repository::ValuesRepository;
 use crate::sub_agent::SubAgentCallbacks;
 use crate::super_agent::config::AgentID;
+use crate::values::values_repository::ValuesRepository;
 use crossbeam::channel::never;
 use crossbeam::select;
 use opamp_client::StartedClient;
@@ -178,8 +178,8 @@ pub mod test {
     use crate::opamp::remote_config_hash::Hash;
     use crate::sub_agent::error::SubAgentError;
     use crate::sub_agent::event_processor::{EventProcessor, SubAgentEventProcessor};
-    use crate::sub_agent::values::values_repository::test::MockRemoteValuesRepositoryMock;
     use crate::super_agent::config::AgentID;
+    use crate::values::values_repository::test::MockRemoteValuesRepositoryMock;
     use mockall::mock;
     use opamp_client::opamp::proto::RemoteConfigStatus;
     use opamp_client::opamp::proto::RemoteConfigStatuses::Applying;
