+++ super-agent/src/values/k8s.rs
@@ -2,8 +2,8 @@ use crate::agent_type::agent_values::AgentValues;
 use crate::agent_type::definition::AgentType;
 use crate::k8s;
 use crate::k8s::store::{K8sStore, STORE_KEY_LOCAL_DATA_CONFIG, STORE_KEY_OPAMP_DATA_CONFIG};
-use crate::sub_agent::values::values_repository::{ValuesRepository, ValuesRepositoryError};
 use crate::super_agent::config::AgentID;
+use crate::values::values_repository::{ValuesRepository, ValuesRepositoryError};
 use std::sync::Arc;
 use thiserror::Error;
 use tracing::debug;
