+++ super-agent/src/values/values_repository.rs
@@ -62,8 +62,8 @@ pub trait ValuesRepository: Send + Sync + 'static {
 pub mod test {
     use crate::agent_type::agent_values::AgentValues;
     use crate::agent_type::definition::AgentType;
-    use crate::sub_agent::values::values_repository::{ValuesRepository, ValuesRepositoryError};
     use crate::super_agent::config::AgentID;
+    use crate::values::values_repository::{ValuesRepository, ValuesRepositoryError};
     use mockall::{mock, predicate};
 
     mock! {
